Create a Prismic and a Content API variant for StoryPromo
