+++ content/webapp/components/CardGrid/index.tsx
@@ -16,7 +16,6 @@ import ExhibitionGuidePromo from '@weco/content/components/ExhibitionGuidePromo'
 import ExhibitionPromo from '@weco/content/components/ExhibitionPromo';
 import MoreLink from '@weco/content/components/MoreLink';
 import StoryPromo from '@weco/content/components/StoryPromo';
-import StoryPromoContentApi from '@weco/content/components/StoryPromo/StoryPromoContentApi';
 import { Article } from '@weco/content/services/wellcome/content/types/api';
 import { convertItemToCardProps } from '@weco/content/types/card';
 import { Link } from '@weco/content/types/link';
@@ -78,13 +77,18 @@ const CardGrid: FunctionComponent<Props> = ({
                 />
               )}
               {item.type === 'Article' && (
-                <StoryPromoContentApi
+                <StoryPromo
+                  variant="contentApi"
                   article={item}
                   hidePromoText={hidePromoText}
                 />
               )}
               {item.type === 'articles' && (
-                <StoryPromo article={item} hidePromoText={hidePromoText} />
+                <StoryPromo
+                  variant="prismic"
+                  article={item}
+                  hidePromoText={hidePromoText}
+                />
               )}
               {(item.type === 'exhibition-guides' ||
                 item.type === 'exhibition-highlight-tours' ||
